backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-dotenv
pillow